% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case}
\alias{case}
\title{Genotypes for the affected children of case-parent triads.}
\format{
A data frame with 1000 rows and 100 variables
}
\usage{
data(case)
}
\description{
A simulated dataset containing the counts of the alternate allele
for 100 SNPs for the affected child in 1000 simulated case-parent
triads. Columns represent SNPs, rows are individuals. SNPs
in columns 51, 52, 76, and 77 represent a true risk pathway.
}
\keyword{datasets}
