% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.best.weights.R
\name{choose.best.weights}
\alias{choose.best.weights}
\title{A function to identify the best family weights for GADGETS and re-run islands that did not use those}
\usage{
choose.best.weights(results.dir, preprocessed.list)
}
\arguments{
\item{results.dir}{The directory in which individual island results from \code{run.gadgets} are saved.}

\item{preprocessed.list}{The initial list produced by function \code{preprocess.genetic.data}.}
}
\value{
A data.table containing the results aggregated across islands. Note these results be written to \code{results.dir}
as 'combined.island.unique.chromosome.results.rds'. See the package vignette for more detailed descriptions of the content
of each output column.
}
\description{
This function chooses the best family weights among those input for GADGETS and re-run islands that did not
use those best weights.
}
\examples{

data(case)
data(dad)
data(mom)
data(snp.annotations)

pp.list <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               ld.block.vec = c(10))

run.gadgets(pp.list, n.chromosomes = 4, chromosome.size = 3, results.dir = 'tmp',
       cluster.type = 'interactive', registryargs = list(file.dir = 'tmp_reg', seed = 1500),
       generations = 2, n.islands = 2, island.cluster.size = 1,
       n.migrations = 0)

best.weights <- choose.best.weights('tmp', pp.list)

unlink('tmp', recursive = TRUE)
unlink('tmp_reg', recursive = TRUE)

}
