% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.fitness.score.R
\name{chrom.fitness.score}
\alias{chrom.fitness.score}
\title{A function to generate a fitness score for a subset of snps from a dataset of case genetic markers}
\usage{
chrom.fitness.score(
  case.comp.differences,
  target.snps,
  cases.minus.complements,
  both.one.mat,
  n.different.snps.weight = 2,
  n.both.one.weight = 1
)
}
\arguments{
\item{case.comp.differences}{a data frame or matrix indicating case genetic data != complement genetic data, where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the snps for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to case genetic data - complement genetic data.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the alternate allele, equal to (case.genetic.data == 1 & complement.genetic.data == 1).}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}
}
\value{
A list whose first element is the fitness score and second element is the sum of weighted difference vectors for the target snps.
}
\description{
This function returns a fitness score for a set of snps, where the fitness score is the squared vector length of the weighted sum of difference vectors between cases and complements.
}
\examples{

data(case)
data(dad)
data(mom)
comp <- mom + dad - case
case.comp.diff <- case != comp
case.minus.comp <- case - comp
both.one.mat <- case == 1 & comp == 1
chrom.fitness.score(case.comp.diff, target.snps = c(1, 4, 7), case.minus.comp, both.one.mat)

}
