% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.islands.R
\name{combine.islands}
\alias{combine.islands}
\title{A function to combine results for individual islands into a single dataset.}
\usage{
combine.islands(results.dir)
}
\arguments{
\item{results.dir}{The directory in which individual island results from \code{run.ga} are saved.}
}
\value{
A data.table combing results from all islands stored in \code{results.dir}. This will also be written to \code{results.dir}
}
\description{
This function combines results for individual islands into a single dataset.
}
