% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.epi.fitness.scores.R
\name{compute.epi.fitness.scores}
\alias{compute.epi.fitness.scores}
\title{A function to combine results for individual islands into a single dataset.}
\usage{
compute.epi.fitness.scores(
  results.dir,
  preprocessed.list,
  n.different.snps.weight = 2,
  n.both.risk.geno.weight = 1,
  weight.function.int = 2
)
}
\arguments{
\item{results.dir}{The directory in which individual island results from \code{run.ga} are saved.}

\item{preprocessed.list}{The initial list produced by function \code{preprocess.genetic.data}.}

\item{n.different.snps.weight}{The number by which the number of different SNPs between a case and complement is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.risk.geno.weight}{The number by which the number of SNPs with the risk genotype in both the case and complement is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function.int}{An integer used to assign family weights. Specifically, we use \code{weight.function.int} in a  function that takes the weighted sum
of the number of different SNPs and SNPs both equal to one as an argument, denoted as x, and returns a family weight equal to \code{weight.function.int}^x. Defaults to 2.}
}
\value{
A list of two elements. Note these two objects will also be written to \code{results.dir}
as 'epi.combined.island.results.rds' and 'epi.combined.island.unique.chromosome.results.rds'. Furthermore,
this function should not be called twice on the same directory.
\describe{
 \item{epi.all.results}{A dataset containing chromosome results across all islands,
 where top chromosomes that evolved on multiple distinct islands appear in multiple rows.}
 \item{epi.unique.results}{A condensed version of \code{all.results} with one row per distinct chromosome
 and an additional variable indicating the number of islands on which that chromosome evolved.}
}
}
\description{
This function combines results for individual islands into a single dataset. It can only be run after calling function \code{combine.islands}.
}
\examples{

data(case)
data(dad)
data(mom)
data(snp.annotations)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))

pp.list <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               chrom.mat = chrom.mat[ , 1:10])

run.ga(pp.list, n.chromosomes = 4, chromosome.size = 3, results.dir = 'tmp',
       cluster.type = 'interactive', registryargs = list(file.dir = 'tmp_reg', seed = 1500),
       generations = 2, n.islands = 2, island.cluster.size = 1, n.top.chroms = 3)

combined.res <- combine.islands('tmp', snp.annotations[ 1:10, ], pp.list)
epi.score.res <- compute.epi.fitness.scores('tmp', pp.list)

unlink('tmp', recursive = TRUE)
unlink('tmp_reg', recursive = TRUE)

}
