% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dad}
\alias{dad}
\title{SNP counts for the fathers of case-parent triads.}
\format{
A data frame with 1000 rows and 100 variables
}
\usage{
data(dad)
}
\description{
A simulated dataset containing the counts of the alternate allele
for 100 SNPs for the fathers in 1000 simulated case-parent
triads. Columns represent SNPs, rows are individuals. SNPs
in columns 51, 52, 76, and 77 represent a true risk pathway.
}
\keyword{datasets}
