% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.chrom.fitness.score.R
\name{epi.chrom.fitness.score}
\alias{epi.chrom.fitness.score}
\title{A function to assign a fitness score to a chromosome using epistasis weighting}
\usage{
epi.chrom.fitness.score(
  case.genetic.data,
  complement.genetic.data,
  target.snps,
  target.snps.sum.diff.vecs,
  target.snps.risk.alleles,
  chrom.mat,
  weight.lookup,
  n.different.snps.weight = 2,
  n.both.risk.geno.weight = 1,
  n.case.high.risk.thresh = 20,
  outlier.sd = 2.5,
  epi.test = FALSE
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios. Columns are SNPs, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts.
Columns are SNPs, rows are families.}

\item{target.snps}{A numeric vector of the columns corresponding to the collection of SNPs, or chromosome, for which the fitness score will be computed.}

\item{target.snps.risk.alleles}{A corresponding to the \code{risk.set.alleles} output from \code{chrom.fitness.score} for
\code{target.snps}.}

\item{chrom.mat}{A logical matrix indicating whether the SNPs in \code{case.comp.differences} are located on the same biological chromosome.}

\item{weight.lookup}{A vector that maps a family weight to the weighted sum of the number of different SNPs and SNPs both equal to one.}

\item{n.different.snps.weight}{The number by which the number of different SNPs between a case and complement is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.risk.geno.weight}{The number by which the number of SNPs with the risk genotype in both the case and complement is multiplied in computing the family weights. Defaults to 1.}

\item{n.case.high.risk.thresh}{The number of cases with the provisional high risk set required to check for recessive patterns of allele inheritance.}

\item{outlier.sd}{The number of standard deviations from the mean allele count used to determine whether recessive allele coding is appropriate}

\item{epi.test}{A logical indicating whether the function should return the information required to run function \code{epi.test}.
for a given SNP. See the GADGET paper for specific details on the implementation of this argument.}

\item{case.comp.differences}{A data frame or matrix indicating \code{case.genetic.data} != \code{complement.genetic.data},
where rows correspond to individuals and columns correspond to snps.}

\item{cases.minus.complements}{A matrix equal to \code{case.genetic.data} - \code{complement genetic data}.}

\item{target.snps.sum.dif.vecs}{A vector corresponding to the \code{sum.dif.vecs} output from \code{chrom.fitness.score} for
\code{target.snps}.}
}
\value{
A list:
\describe{
 \item{fitness.score}{The chromosome fitness score.}
 \item{sum.dif.vecs}{The weighted mean difference vector corresponding to the chromosome.
 The magnitudes of these values are not particularly important, but the sign is useful.
 A positive value for a given SNP indicates the minor allele is positively associated with
 disease status, while a negative value implies the reference (‘wild type’) allele is
 positively associated with the disease.}
 \item{rr}{The fraction of provisional risk alleles carried by cases with the full risk set
 over the total number of risk alleles carried by either a case or complement with the full risk set.}
 \item{pseudo.t2}{The pseudo T^2 value for the chromosome.}
 \item{risk.set.alleles}{A vector indicating the number risk alleles a case or complement must have
  for each SNP in \code{target.snps} for the case or complement to be classified as having the
  proposed risk set. '1+' indicates at least one copy of the risk allele is required, while '2'
  indicates 2 copies are needed. The risk allele can be determined based on the signs of the elements
  of \code{sum.dif.vecs}, where a negative value indicates the major allele for a given SNP is
  the risk allele, while a positive value implicates the minor allele.}
}
}
\description{
This function assigns a fitness score to a chromosome using epistasis weighting.
}
\examples{

data(case)
data(dad)
data(mom)
comp <- mom + dad - case
case.comp.diff <- case != comp
case.minus.comp <- sign(case - comp)
both.one.mat <- case == 1 & comp == 1
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))
weight.lookup <- vapply(seq_len(6), function(x) 2^x, 1)
first.pass <- chrom.fitness.score(case, comp, case.comp.diff, c(1, 4, 7),
                            case.minus.comp, both.one.mat,
                            chrom.mat, weight.lookup)
epi.chrom.fitness.score(case, comp, case.comp.diff, c(1, 4, 7), case.minus.comp,
                        first.pass$sum.dif.vecs, first.pass$risk.set.alleles,
                        chrom.mat, weight.lookup)

}
