% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistasisGA.R
\docType{package}
\name{epistasisGAGE}
\alias{epistasisGAGE}
\title{\code{epistasisGA} package}
\description{
A package implementing the GADGETS method to detect multi-SNP effects in
case-parent triad or affected/unaffected sibling studies.
}
