% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure}
\alias{exposure}
\title{Exposures for the cases of case-parent triads with a simulated gene
environment interaction.}
\format{
A data frame with 1000 rows and 100 variables
}
\usage{
data(exposure)
}
\description{
A simulated vector containing the exposure status for each case
of the case-parent triads data.
.
}
\keyword{datasets}
