% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.score.cpp.R
\name{fitness.score.cpp}
\alias{fitness.score.cpp}
\title{A function to generate a fitness score for a subset of snps from a dataset of case genetic markers}
\usage{
fitness.score.cpp(case.genetic.data, target.snps, k = 10,
  correct.thresh = 0.9)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{target.snps}{A numeric vector of the columns corresponding to the snps for which the fitness score will be computed.}

\item{k}{A numeric scalar corresponding to the number of nearest neighbors required for computing the fitness score. See details for more information.}

\item{correct.thresh}{A numeric scalar between 0 and 1 indicating the minimum proportion of of cases among the nearest neighbors for a given individual for that individual to be considered correctly classified. See details for more information.}
}
\value{
A scalar, the fitness score for the given set of snps.
}
\description{
This function takes a subset of snps from an input genetic dataset and returns a fitness score using a k-nearest neighbors approach.
}
\examples{

data(cases)
fitness.score.cpp(target.snps = c(1, 4, 7), case.genetic.data = cases)

}
