% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{ga}
\alias{ga}
\title{A function to run a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.}
\usage{
ga(
  case.genetic.data,
  father.genetic.data,
  mother.genetic.data,
  n.chromosomes,
  chromosome.size,
  dist.type,
  generations = 2000,
  k = 10,
  correct.thresh = 0.9,
  gen.same.fitness = 500,
  tol = 10^-6,
  n.top.chroms = 100,
  run.parallel = F
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{father.genetic.data}{The genetic data for the father of the case. Columns are snps, rows are individuals.}

\item{mother.genetic.data}{The genetic data for the mother of the case. Columns are snps, rows are individuals.}

\item{n.chromosomes}{A scalar indicating the number of candidate collections of snps to use in the GA.}

\item{chromosome.size}{The number of snps within each candidate solution.}

\item{dist.type}{A character string indicating the type of distance measurement. Type 'knn' performs k-nearest neighbors classifications, type 'paired' performs paired length classifications.}

\item{generations}{The maximum number of generations for which the GA will run. Defaults to 2000.}

\item{k}{A numeric scalar corresponding to the number of nearest neighbors required for computing the fitness score. Defaults to 10}

\item{correct.thresh}{A numeric scalar between 0 and 1 indicating the minimum proportion of of cases among the nearest neighbors for a given individual for that individual to be considered correctly classified. Defaults to 0.9.}

\item{gen.same.fitness}{The number of consecutive generations with the same fitness score required for algorithm termination.}

\item{tol}{The maximum absolute pairwise difference among the top fitness scores from the previous 500 generations considered to be sufficient to stop producing new generations.}

\item{n.top.chroms}{The number of top scoring chromosomes, according to fitness score, to return.}

\item{run.parallel}{Logical indicator of whether the fitness scores should be computed in parallel using bplapply from bioconductor. If FALSE, will use sapply instead. Defaults to FALSE.}
}
\value{
A list, whose first element is a list of the top 100 scoring chromosomes, and second element is a vector of the corresponding fitness scores.
}
\description{
This function runs a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.
}
\examples{

data(case)
data(dad)
data(mom)

ga.res <- ga(case, dad, mom, 7, 3, dist.type = 'knn', generations = 1, k = 10,
 correct.thresh = 0.9, tol = 10^-6, run.parallel = TRUE)

}
