% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mom.mci}
\alias{mom.mci}
\title{Genotypes for the mothers of case-parent triads with a simulated
maternal-fetal interaction.}
\format{
A matrix with 1000 rows and 24 variables
}
\usage{
data(mom.mci)
}
\description{
A simulated dataset containing the counts of the alternate allele
for 24 SNPs for the mothers in 1000 simulated case-parent
triads. Columns represent SNPs, rows are individuals. The SNP
in column 6 interacts with the SNPs in columns 12 and 18
of dataset \code{case.mci} to increase risk of disease in the child,
where at least one copy of the alternate allele (genotype 1 or 2)
is required at each implicated locus.
.
}
\keyword{datasets}
