% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mom}
\alias{mom}
\title{SNP counts for the mothers of case-parent triads.}
\format{
A data frame with 1000 rows and 100 variables
}
\usage{
mom
}
\description{
A dataset containing the counts of the alternate allele
for 100 SNPs for the mothers in 1000 simulated case-parent
triads. Columns represent snps, rows are individuals.
}
\keyword{datasets}
