% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot.R
\name{network.plot}
\alias{network.plot}
\title{A function to plot a network of potential epistatic snps}
\usage{
network.plot(
  results.df,
  repulse.rad = 1000,
  node.size = 25,
  graph.area = 100,
  vertex.label.cex = 0.5,
  seed = 10,
  plot = TRUE
)
}
\arguments{
\item{results.df}{The \code{unique.results} data frame of results of GA runs from \code{combine.islands}, with fitness scores restricted to the top results using function \code{network.threshold}.}

\item{repulse.rad}{A scalar affecting the graph shape. Decrease to reduce overlapping nodes.}

\item{node.size}{A scalar affecting the size of the graph nodes. Increse to increase size.}

\item{graph.area}{A scalar affecting the size of the graph area. Increase to increase graph area.}

\item{vertex.label.cex}{A scalar controlling the size of the vertex label. Increase to increase size.}

\item{seed}{an integer specifying the graph seed}

\item{plot}{A logical indicating whether the network should be plotted. If set to false, this function will return an igraph object which can be used for manual plotting.}
}
\value{
An igraph object, if \code{plot} is set to FALSE.
}
\description{
This function rplot a network of potential epistatic snps.
}
