% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.threshold.R
\name{network.threshold}
\alias{network.threshold}
\title{A function to identify a threshold fitness score above which chromosomes should be included in a network plot}
\usage{
network.threshold(observed.results, permutation.list, threshold.val)
}
\arguments{
\item{observed.results}{A data frame of results of GA runs for the observed data, containing only unique chromosomes. The first n columns are snp ids, the next n columns are there associated mean vector components.}

\item{permutation.list}{A list of data frames, each containing unique results for GA runs, for permutations of the observed data.}

\item{threshold.val}{The desired theshold. This function will return a fitness score such that \code{threshold.val}*100\%  of permutations have at least one chromosome exceeding this score.}
}
\value{
A numeric fitness score such that \code{threshold.val}*100\%  of permutations have at least one chromosome exceeding this score.
}
\description{
This function identifies a threshold fitness score above which chromosomes should be included in a network plot
}
