% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.genetic.data.R
\name{preprocess.genetic.data}
\alias{preprocess.genetic.data}
\title{A function to code the genetic data according to specific models of inheritance identify using a data driven approach.}
\usage{
preprocess.genetic.data(
  case.genetic.data,
  complement.genetic.data = NULL,
  father.genetic.data = NULL,
  mother.genetic.data = NULL,
  chrom.mat,
  min.allele.freq = 0.025,
  bp.param = bpparam()
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where \code{complement.genetic.data} = mother snp counts + father snp counts - case snp counts. Columns are snps, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{father.genetic.data}{The genetic data for the father of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{mother.genetic.data}{The genetic data for the mother of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{chrom.mat}{A logical matrix indicating whether the snps in the input genetic data belong to the same chromosome.}

\item{min.allele.freq}{The minimum minor allele frequency in the parents required for a snp to be considered as a potential GA solution. Any snps with MAF < \code{min.allele.freq} in the parents will be omitted. Defaults to 0.01.}

\item{bp.param}{The BPPARAM argument to be passed to bplapply when running conditional logistic regressions on each input SNP. If using a cluster computer, this parameter needs to be set with care.}
}
\value{
A list, whose first element is a matrix of case genetic data, second element is complment genetic data, and third element is a vector of chi-square statistics, fourth element is a list of column numbers in the original dataset to whom the .
}
\description{
This function performs model selection on each input snp and recodes the data based on the results.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500))))
res <- preprocess.genetic.data(case[, 1:3], father.genetic.data = dad[ , 1:3], mother.genetic.data = mom[ , 1:3],
                 chrom.mat = chrom.mat[ , 1:3])

}
