% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ga.R
\name{run.ga}
\alias{run.ga}
\title{A function to run a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.}
\usage{
run.ga(
  case.genetic.data,
  father.genetic.data,
  mother.genetic.data,
  n.chromosomes,
  chromosome.size,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  generations = 2000,
  gen.same.fitness = 500,
  tol = 10^-6,
  n.top.chroms = 100
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{father.genetic.data}{The genetic data for the father of the case. Columns are snps, rows are individuals.}

\item{mother.genetic.data}{The genetic data for the mother of the case. Columns are snps, rows are individuals.}

\item{n.chromosomes}{A scalar indicating the number of candidate collections of snps to use in the GA.}

\item{chromosome.size}{The number of snps within each candidate solution.}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number different snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}

\item{generations}{The maximum number of generations for which the GA will run. Defaults to 2000.}

\item{gen.same.fitness}{The number of consecutive generations with the same fitness score required for algorithm termination.}

\item{tol}{The maximum absolute pairwise difference among the top fitness scores from the previous 500 generations considered to be sufficient to stop producing new generations.}

\item{n.top.chroms}{The number of top scoring chromosomes, according to fitness score, to return.}
}
\value{
A list, whose first element is a list of the top 100 scoring chromosomes, and second element is a vector of the corresponding fitness scores.
}
\description{
This function runs a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.
}
\examples{

data(case)
data(dad)
data(mom)

ga.res <- run.ga(case, dad, mom, 7, 3, generations = 1)

}
