% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.global.test.R
\name{run.global.test}
\alias{run.global.test}
\title{A function to run a global test for significance of top chromosome results across a range of chromosome sizes}
\usage{
run.global.test(results.list)
}
\arguments{
\item{results.list}{A list of length d, where d is the number of chromosome sizes to be included in a global test.
Each element of the list must itself be a list whose first element \code{observed.data} is a data.table containing
the unique chromosome results from \code{combine.islands} for a given chromosome size. The second element \code{permutation.list}
is a list containing all permutation results data.tables, again using the unique chromosome results output by \code{combine.islands}
for each permutation.}
}
\value{
A list containing the observed test statistic \code{obs.test.stat}, the p-value \code{pval},
and a vector of permutation test statistics \code{perm.test.stats}.
}
\description{
This function runs a global test for significance of top chromosome results across a range of chromosome sizes.
}
