% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp.annotations.mci}
\alias{snp.annotations.mci}
\title{RSID, REF, and ALT annotations for example dataset SNPs}
\format{
A matrix with 24 rows and 3 variables
}
\usage{
data(snp.annotations.mci)
}
\description{
A data.frame containing the RSID, REF allele and ALT allele
for each SNP in the example datasets \code{case.mci}, \code{mom.mci},
\code{dad.mci}, \code{case.gxe}, \code{mom.gxe}, and \code{dad.gxe}.
The SNPs are in the same order as they appear in the example dataset columns.
}
\keyword{datasets}
