% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp.annotations}
\alias{snp.annotations}
\title{RSID, REF, and ALT annotations for example dataset SNPs}
\format{
A data frame with 100 rows and 3 variables
}
\usage{
data(snp.annotations)
}
\description{
A data.frame containing the RSID, REF allele and ALT allele
for each SNP in the example datasets. The SNPs are in the same
order as they appear in the example dataset columns.
}
\keyword{datasets}
