% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.fitness.score.R
\name{chrom.fitness.score}
\alias{chrom.fitness.score}
\title{A function to generate a fitness score for a subset of snps from a dataset of case genetic markers}
\usage{
chrom.fitness.score(
  case.genetic.data,
  complement.genetic.data,
  case.comp.differences,
  target.snps,
  cases.minus.complements,
  both.one.mat,
  chrom.mat,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  weight.function = function(x) 2^x,
  n.case.high.risk.thresh = 20
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where \code{complement.genetic.data} = mother snp counts + father snp counts - case snp counts. Columns are snps, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{case.comp.differences}{a data frame or matrix indicating case genetic data != complement genetic data, where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the snps for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to case genetic data - complement genetic data.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the alternate allele, equal to (case.genetic.data == 1 & complement.genetic.data == 1).}

\item{chrom.mat}{A logical matrix indicating whether the snps in \code{case.comp.differences} belong to the same chromosome.}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function}{A function which takes the weighted sum of the number of different snps and snps both equal to one as an argument, and returns a family weight. Defaults to the identity function.}

\item{n.case.high.risk.thresh}{The number of cases with the provisional high risk set required to check for recessive patterns of allele inheritance.}
}
\value{
A list whose first element is the fitness score and second element is the sum of weighted difference vectors for the target snps.
}
\description{
This function returns a fitness score for a set of snps, where the fitness score is the squared vector length of the weighted sum of difference vectors between cases and complements.
}
\examples{

data(case)
data(dad)
data(mom)
comp <- mom + dad - case
case.comp.diff <- case != comp
case.minus.comp <- case - comp
both.one.mat <- case == 1 & comp == 1
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))
chrom.fitness.score(case, comp, case.comp.diff, c(1, 4, 7),
                    case.minus.comp, both.one.mat, chrom.mat)

}
