% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom.fitness.score.R
\name{chrom.fitness.score}
\alias{chrom.fitness.score}
\title{A function to assign a fitness score to a chromosome}
\usage{
chrom.fitness.score(
  case.genetic.data,
  complement.genetic.data,
  case.comp.differences,
  target.snps,
  cases.minus.complements,
  both.one.mat,
  chrom.mat,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  weight.function = function(x) 2^x,
  n.case.high.risk.thresh = 20
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios. Columns are SNPs, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts.
Columns are SNPs, rows are families.}

\item{case.comp.differences}{A data frame or matrix indicating \code{case.genetic.data} != \code{complement.genetic.data},
where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the collection of SNPs, or chromosome, for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to \code{case.genetic.data} - \code{complement genetic data}.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the minor allele,
equal to \code{case.genetic.data == 1 & complement.genetic.data == 1}.}

\item{chrom.mat}{A logical matrix indicating whether the SNPs in \code{case.comp.differences} are located on the same biological chromosome.}

\item{n.different.snps.weight}{The number by which the number of different SNPs between a case and complement is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of SNPs equal to 1 in both the case and complement is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function}{A function that takes the weighted sum of the number of different SNPs and SNPs both equal to one as an argument, denoted as x,
and returns a family weight. Defaults to 2^x.}

\item{n.case.high.risk.thresh}{The number of cases with the provisional high risk set required to check for recessive patterns of allele inheritance.}
}
\value{
A list:
\describe{
 \item{fitness.score}{The chromosome fitness score.}
 \item{sum.dif.vecs}{The weighted mean difference vector corresponding to the chromosome.}
 \item{rr}{The fraction of provisional risk alleles carried by cases with the full risk set
 over the total number of risk alleles carried by either a case or complement with the full risk set.}
 \item{pseudo.t2}{The pseudo T^2 value for the chromosome.}
}
}
\description{
This function assigns a fitness score to a chromosome.
}
\examples{

data(case)
data(dad)
data(mom)
comp <- mom + dad - case
case.comp.diff <- case != comp
case.minus.comp <- case - comp
both.one.mat <- case == 1 & comp == 1
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))
chrom.fitness.score(case, comp, case.comp.diff, c(1, 4, 7),
                    case.minus.comp, both.one.mat, chrom.mat)

}
