% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.islands.R
\name{combine.islands}
\alias{combine.islands}
\title{A function to combine results for individual islands into a single dataset.}
\usage{
combine.islands(results.dir)
}
\arguments{
\item{results.dir}{The directory in which individual island results from \code{run.ga} are saved.}
}
\value{
A list of two elements, \code{all.results} and \code{unique.results}. \code{all.results} contains chromosome results across all islands, including chromsomes identified on more than one island. \code{unique.results} removes duplicate chromosomes found on more than one island. These two objects will also be written to \code{results.dir} as combined.island.results.rds and combined.island.unique.chromosome.results.rds.
}
\description{
This function combines results for individual islands into a single dataset.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500))))

pp.list <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               chrom.mat = chrom.mat[ , 1:10])

run.ga(pp.list, n.chromosomes = 4, chromosome.size = 3, results.dir = "tmp",
       cluster.type = "interactive", registryargs = list(file.dir = "tmp_reg", seed = 1500),
       generations = 2, n.islands = 2, island.cluster.size = 1, n.top.chroms = 3)

combined.res <- combine.islands("tmp")

}
