% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.islands.R
\name{combine.islands}
\alias{combine.islands}
\title{A function to combine results for individual islands into a single dataset.}
\usage{
combine.islands(results.dir)
}
\arguments{
\item{results.dir}{The directory in which individual island results from \code{run.ga} are saved.}
}
\value{
A list of two elements. Note these two objects will also be written to \code{results.dir}
as 'combined.island.results.rds' and 'combined.island.unique.chromosome.results.rds'.
\describe{
 \item{all.results}{A dataset containing chromosome results across all islands,
 where top chromosomes that evolved on multiple distinct islands appear in multiple rows.}
 \item{unique.results}{A condensed version of \code{all.results} with one row per distinct chromosome
 and an additional variable indicating the number of islands on which that chromosome evolved.}
}
}
\description{
This function combines results for individual islands into a single dataset.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))

pp.list <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               chrom.mat = chrom.mat[ , 1:10])

run.ga(pp.list, n.chromosomes = 4, chromosome.size = 3, results.dir = 'tmp',
       cluster.type = 'interactive', registryargs = list(file.dir = 'tmp_reg', seed = 1500),
       generations = 2, n.islands = 2, island.cluster.size = 1, n.top.chroms = 3)

combined.res <- combine.islands('tmp')

unlink('tmp', recursive = TRUE)
unlink('tmp_reg', recursive = TRUE)

}
