% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.island.R
\name{evolve.island}
\alias{evolve.island}
\title{A function to evolve a genetic algorithm for a specific island for a given number of generations}
\usage{
evolve.island(
  n.migrations,
  case.genetic.data,
  complement.genetic.data,
  case.comp.different,
  case.minus.comp,
  both.one.mat,
  chrom.mat,
  n.chromosomes,
  n.candidate.snps,
  chromosome.size,
  start.generation,
  seed.val,
  snp.chisq,
  original.col.numbers,
  all.converged = F,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  weight.function = identity,
  total.generations = 50,
  gen.same.fitness = 50,
  max.generations = 500,
  tol = 10^-6,
  n.top.chroms = 100,
  initial.sample.duplicates = F,
  snp.sampling.type = "chisq",
  crossover.prop = 0.8,
  chromosome.list = NULL,
  fitness.score.mat = NULL,
  top.fitness = NULL,
  last.gens.equal = NULL,
  top.generation.chromosome = NULL,
  chromosome.mat.list = NULL,
  sum.dif.vec.list = NULL,
  n.case.high.risk.thresh = 20
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where \code{complement.genetic.data} = mother snp counts + father snp counts - case snp counts. Columns are snps, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the alternate allele, equal to (case.genetic.data == 1 & complement.genetic.data == 1).}

\item{chrom.mat}{A logical matrix indicating whether the snps in \code{case.comp.differences} belong to the same chromosome.}

\item{n.chromosomes}{A scalar indicating the number of candidate collections of snps to use in the GA.}

\item{n.candidate.snps}{A scalar indicating the number eligible snps in the input data, after filtering out low MAF SNPs.}

\item{chromosome.size}{The number of snps within each candidate solution.}

\item{start.generation}{The generation at which this function should begin. If 1, a random set of chromosomes will be initialized. Otherwise the argument \code{chromosome.list} will be used.}

\item{seed.val}{An integer indicating the seed to be used for the random samples.}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of different snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function}{A function that takes the weighted sum of the number of different snps and snps both equal to one as an argument, and returns a family weight. Defaults to the identity function.}

\item{total.generations}{The number of generations for which the GA will run. Defaults to 50.}

\item{gen.same.fitness}{The number of consecutive generations with the same fitness score required for algorithm termination.}

\item{max.generations}{The total allowable number of generations. Defaults to 500.}

\item{tol}{The maximum absolute pairwise difference among the top fitness scores from the previous 500 generations considered to be sufficient to stop producing new generations.}

\item{n.top.chroms}{The number of top scoring chromosomes, according to fitness score, to return.}

\item{initial.sample.duplicates}{A logical indicating whether the same snp can appear in more than one chromosome in the initial sample of chromosomes (the same snp may appear in more than one chromosome thereafter, regardless). Default to F.}

\item{snp.sampling.type}{A string indicating how snps are to be sampled for mutations. Options are "zscore" or "random". Defaults to "zscore".}

\item{crossover.prop}{A numeric between 0 and 1 indicating the proportion of chromosomes to be subjected to cross over. The remaining proportion will be mutated. Defaults to 0.8.}

\item{chromosome.list}{A list of chromosomes on which the genetic algorithm will start}

\item{n.case.high.risk.thresh}{The number of cases with the provisional high risk set required to check for recessive patterns of allele inheritance.}

\item{case.comp.differences}{a data frame or matrix indicating case genetic data != complement genetic data, where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the snps for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to case genetic data - complement genetic data.}
}
\value{
A list, whose first element is a data.table of the top \code{n.top.chroms scoring chromosomes}, their fitness scores, and their difference vectors. The second element is a scalar indicating the number of generations required to identify a solution.
}
\description{
This function evolves a genetic algorithm for a given number of generations.
}
