% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposure}
\alias{exposure}
\title{Exposures for the cases of case-parent triads with a simulated gene
environment interaction.}
\format{
A data frame with 1000 rows and 1 variables
}
\usage{
data(exposure)
}
\description{
A data.frame containing simulated exposure status for each case
of the case-parent triads data. Rows correspond to different
families. The single column represents a binary exposure, where
in combination with the relevant risk-associated alleles (columns
6, 12, and 18 in data set case.gxe), is associated with increased risk.
.
}
\keyword{datasets}
