% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxE.fitness.score.R
\name{GxE.fitness.score}
\alias{GxE.fitness.score}
\title{A function to assign a fitness score to a chromosome}
\usage{
GxE.fitness.score(
  case.genetic.data,
  complement.genetic.data,
  case.comp.differences,
  target.snps,
  cases.minus.complements,
  both.one.mat,
  ld.block.vec,
  weight.lookup,
  case2.mat,
  case0.mat,
  exposure,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  recode.threshold = 3
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios or affected/unaffected sibling pairs.
 Columns are SNP allele counts, and rows are individuals.
The ordering of the columns must be consistent with the LD structure specified in \code{ld.block.vec}.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts.
Columns are SNP allele counts, rows are families. If using affected/unaffected sibling pairs, this should contain
the unaffected sibling genotypes.}

\item{case.comp.differences}{A data frame or matrix indicating \code{case.genetic.data} != \code{complement.genetic.data},
where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the collection of SNPs, or chromosome, for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to \code{case.genetic.data} - \code{complement genetic data}.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the minor allele,
equal to \code{case.genetic.data == 1 & complement.genetic.data == 1}.}

\item{ld.block.vec}{An integer vector specifying the linkage blocks of the input SNPs. As an example, for 100 candidate SNPs, suppose
we specify \code{ld.block.vec <- c(25, 75, 100)}. This vector indicates that the input genetic data has 3 distinct linkage blocks, with
SNPs 1-25 in the first linkage block, 26-75 in the second block, and 76-100 in the third block. Note that this means the ordering of the columns (SNPs)
in \code{case.genetic.data} must be consistent with the LD blocks specified in \code{ld.block.vec}. In the absence of outside information,
a reasonable default is to consider SNPs to be in LD if they are located on the same biological chromosome. If not specified, this defaults
to assuming all input SNPs are in linkage, which may be overly conservative and could adversely affect performance.}

\item{weight.lookup}{A vector that maps a family weight to the weighted sum of the number of different SNPs and SNPs both equal to one.}

\item{case2.mat}{A logical matrix indicating whether, for each SNP, the case carries 2 copies of the minor allele.}

\item{case0.mat}{A logical matrix indicating whether, for each SNP, the case carries 0 copies of the minor allele.}

\item{exposure}{A categorical factor vector corresponding to environmental exposures of the cases.}

\item{n.different.snps.weight}{The number by which the number of different SNPs between a case and complement/unaffected sibling
is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of SNPs equal to 1 in both the case and complement/unaffected sibling
is multiplied in computing the family weights. Defaults to 1.}

\item{recode.threshold}{For a given SNP, the minimum test statistic required to recode and recompute the fitness score using recessive coding. Defaults to 3.
See the GADGETS paper for specific details.}
}
\value{
A list:
\describe{
 \item{fitness.score}{The chromosome fitness score.}
 \item{sum.dif.vecs}{The weighted mean difference vector corresponding to the chromosome,
 with each element divided by it's pseudo-standard error.
 The magnitudes of these values are not particularly important, but the sign is useful.
 A positive value for a given SNP indicates the minor allele is positively associated with
 disease status, while a negative value implies the reference (‘wild type’) allele is
 positively associated with the disease.}
}
}
\description{
This function assigns a fitness score to a chromosome.
}
\examples{

data(case.gxe)
data(dad.gxe)
data(mom.gxe)
data(exposure)
comp.gxe <- mom.gxe + dad.gxe - case.gxe
case.comp.diff <- case.gxe != comp.gxe
case.minus.comp <- case.gxe - comp.gxe
both.one.mat <- case.gxe == 1 & comp.gxe == 1
case2.mat <- case.gxe == 2
case0.mat <- case.gxe == 0
comp2.mat <- comp.gxe == 2
comp0.mat <- comp.gxe == 0
ld.block.vec <- cumsum(rep(25, 4))
weight.lookup <- vapply(seq_len(6), function(x) 2^x, 1)

}
