% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxE.fitness.score.R
\name{GxE.fitness.score}
\alias{GxE.fitness.score}
\title{A function to assign a fitness score to a chromosome}
\usage{
GxE.fitness.score(
  case.genetic.data,
  complement.genetic.data,
  case.comp.differences,
  target.snps,
  cases.minus.complements,
  both.one.mat,
  block.ld.mat,
  weight.lookup,
  case2.mat,
  case0.mat,
  exposure,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  recode.threshold = 3
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios or affected/unaffected sibling pairs.
 Columns are SNP allele counts, and rows are individuals.
The ordering of the columns must be consistent with the LD structure specified in \code{block.ld.mat}.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts.
Columns are SNP allele counts, rows are families. If using affected/unaffected sibling pairs, this should contain
the unaffected sibling genotypes.}

\item{case.comp.differences}{A data frame or matrix indicating \code{case.genetic.data} != \code{complement.genetic.data},
where rows correspond to individuals and columns correspond to snps.}

\item{target.snps}{A numeric vector of the columns corresponding to the collection of SNPs, or chromosome, for which the fitness score will be computed.}

\item{cases.minus.complements}{A matrix equal to \code{case.genetic.data} - \code{complement genetic data}.}

\item{both.one.mat}{A matrix whose elements indicate whether both the case and complement have one copy of the minor allele,
equal to \code{case.genetic.data == 1 & complement.genetic.data == 1}.}

\item{block.ld.mat}{A logical, block diagonal matrix indicating whether the SNPs in \code{case.genetic.data} should be considered
to be in linkage disequilibrium. Note that this means the ordering of the columns (SNPs) in \code{case.genetic.data} must be consistent
with the LD blocks specified in \code{ld.block.mat}. In the absence of outside information, a reasonable default is to consider SNPs
to be in LD if they are located on the same biological chromosome.}

\item{weight.lookup}{A vector that maps a family weight to the weighted sum of the number of different SNPs and SNPs both equal to one.}

\item{case2.mat}{A logical matrix indicating whether, for each SNP, the case carries 2 copies of the minor allele.}

\item{case0.mat}{A logical matrix indicating whether, for each SNP, the case carries 0 copies of the minor allele.}

\item{exposure}{A categorical factor vector corresponding to environmental exposures of the cases.}

\item{n.different.snps.weight}{The number by which the number of different SNPs between a case and complement/unaffected sibling
is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of SNPs equal to 1 in both the case and complement/unaffected sibling
is multiplied in computing the family weights. Defaults to 1.}

\item{recode.threshold}{For a given SNP, the minimum test statistic required to recode and recompute the fitness score using recessive coding. Defaults to 3.
See the GADGETS paper for specific details.}
}
\value{
A list:
\describe{
 \item{fitness.score}{The chromosome fitness score.}
 \item{sum.dif.vecs}{The weighted mean difference vector corresponding to the chromosome,
 with each element divided by it's pseudo-standard error.
 The magnitudes of these values are not particularly important, but the sign is useful.
 A positive value for a given SNP indicates the minor allele is positively associated with
 disease status, while a negative value implies the reference (‘wild type’) allele is
 positively associated with the disease.}
 \item{q}{The fraction of provisional risk alleles carried by cases with the full risk set
 over the total number of risk alleles carried by either a case or complement with the full risk set.}
 \item{pseudo.t2}{The pseudo T^2 value for the chromosome.}
 \item{risk.set.alleles}{A vector indicating the number risk alleles a case or complement must have
  for each SNP in \code{target.snps} for the case or complement to be classified as having the
  proposed risk set. '1+' indicates at least one copy of the risk allele is required, while '2'
  indicates 2 copies are needed. The risk allele can be determined based on the signs of the elements
  of \code{sum.dif.vecs}, where a negative value indicates the major allele for a given SNP is
  the risk allele, while a positive value implicates the minor allele.}
  \item{inf.families}{An integer vector of the informative family rows. Only returned if \code{epi.test} = TRUE.}
  \item{xbar}{The weighted mean difference sign vector. Only returned if \code{GxE} is TRUE.}
  \item{sigma.hat}{The pseudo covariance matrix of the weighted mean difference sign vector. Only returned if \code{GxE} is TRUE.}
  \item{w}{The sum of the family weights. Only returned if \code{GxE} is TRUE.}
}
}
\description{
This function assigns a fitness score to a chromosome.
}
\examples{

data(case)
data(dad)
data(mom)
comp <- mom + dad - case
case.comp.diff <- case != comp
case.minus.comp <- sign(case - comp)
both.one.mat <- case == 1 & comp == 1
case2.mat <- case == 2
case0.mat <- case == 0
library(Matrix)
block.ld.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))
weight.lookup <- vapply(seq_len(6), function(x) 2^x, 1)
set.seed(11)
exposure <- factor(rbinom(nrow(case), 1, 0.3))
GxE.fitness.score(case, comp, case.comp.diff, c(1, 4, 7),
                    case.minus.comp, both.one.mat,
                    block.ld.mat, weight.lookup,
                    case2.mat, case0.mat, exposure)

}
