% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot.R
\name{network.plot}
\alias{network.plot}
\title{A function to plot a network of potential epistatic snps}
\usage{
network.plot(results.df, chrom.size, seed = 10, n.top.chroms = 50)
}
\arguments{
\item{results.df}{A data frame of results of GA runs. The first n columns are snp ids, the next n columns are there associated mean vector components.}

\item{chrom.size}{An integer indicating the number of snps in the chromosome.}

\item{seed}{an integer specifying the graph seed}

\item{n.top.chroms}{An integer indicating the number of top chromosomes to plot}
}
\value{
A list whose first element is the fitness score and second element is the sum of weighted difference vectors for the target snps.
}
\description{
This function rplot a network of potential epistatic snps.
}
