% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.dataset.R
\name{permute.dataset}
\alias{permute.dataset}
\title{A function to create permutation datasets for permutation based hypothesis testing.}
\usage{
permute.dataset(
  case.genetic.data,
  complement.genetic.data = NULL,
  father.genetic.data = NULL,
  mother.genetic.data = NULL,
  n.permutations = 100,
  seed.val = 1
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where \code{complement.genetic.data} = mother snp counts + father snp counts - case snp counts. Columns are snps, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{father.genetic.data}{The genetic data for the father of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{mother.genetic.data}{The genetic data for the mother of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{n.permutations}{The number of permuted datasets to create.}

\item{seed.val}{The starting seed for the random permutations of the data.}
}
\value{
A list of \code{n.permutations} pairs of case and complement data, where the case complement status has been randomly flipped or not flipped.
}
\description{
This function creates permutation datasets for permutation based hypothesis testing.
}
\examples{

data(case)
data(dad)
data(mom)
perm.data.list <- permute.dataset(case, father.genetic.data = dad, mother.genetic.data = mom,
                 n.permutations = 100)
}
