% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.genetic.data.R
\name{preprocess.genetic.data}
\alias{preprocess.genetic.data}
\title{A function to pre-process case-parent triad of affected/unaffected sibling data.}
\usage{
preprocess.genetic.data(
  case.genetic.data,
  complement.genetic.data = NULL,
  father.genetic.data = NULL,
  mother.genetic.data = NULL,
  ld.block.vec = NULL,
  bp.param = bpparam(),
  snp.sampling.probs = NULL,
  categorical.exposures = NULL,
  continuous.exposures = NULL,
  categorical.exposures.sib = NULL,
  continuous.exposures.sib = NULL,
  use.parents = 1,
  mother.snps = NULL,
  child.snps = NULL
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios or affected/unaffected sibling pairs.
Columns are SNP allele counts, and rows are individuals. This object may either be of class 'matrix' OR of class 'big.matrix'. If of class
'big.matrix' it must be file backed as type 'integer' (see the bigmemory package for more information). The ordering of the columns must be consistent
with the LD structure specified in \code{ld.block.vec}. The genotypes cannot be dosages imputed with uncertainty.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts. If using affected/unaffected siblings
this argument should be the genotypes for the unaffected siblings. This object may either be of class 'matrix' OR of class 'big.matrix'. If of class
'big.matrix' it must be file backed as type 'integer' (see the bigmemory package for more information). Columns are SNP allele counts, rows are
families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.
The genotypes cannot be dosages imputed with uncertainty.}

\item{father.genetic.data}{The genetic data for the fathers of the cases. Columns are SNP allele counts, rows are individuals.
Does not need to be specified if \code{complement.genetic.data} is specified. This object may either be of class 'matrix' OR of class 'big.matrix'. If of class
'big.matrix' it must be file backed as type 'integer' (see the bigmemory package for more information). The genotypes cannot be dosages imputed with
uncertainty.}

\item{mother.genetic.data}{The genetic data for the mothers of the cases. Columns are SNP allele counts, rows are individuals.
Does not need to be specified if \code{complement.genetic.data} is specified. This object may either be of class 'matrix' OR of class 'big.matrix'. If of class
'big.matrix' it must be file backed as type 'integer' (see the bigmemory package for more information). The genotypes cannot be dosages imputed with
uncertainty.}

\item{ld.block.vec}{An integer vector specifying the linkage blocks of the input SNPs. As an example, for 100 candidate SNPs, suppose
we specify \code{ld.block.vec <- c(25, 50, 25)}. This vector indicates that the input genetic data has 3 distinct linkage blocks, with
SNPs 1-25 in the first linkage block, 26-75 in the second block, and 76-100 in the third block. Note that this means the ordering of the columns (SNPs)
in \code{case.genetic.data} must be consistent with the LD blocks specified in \code{ld.block.vec}. In the absence of outside information,
a reasonable default is to consider SNPs to be in LD if they are located on the same biological chromosome. If not specified, this defaults
to assuming all input SNPs are in linkage, which may be overly conservative and could adversely affect performance.}

\item{bp.param}{The BPPARAM argument to be passed to bplapply when estimating marginal disease associations for each SNP.
If using a cluster computer, this parameter needs to be set with care. See \code{BiocParallel::bplapply} for more details}

\item{snp.sampling.probs}{A vector indicating the sampling probabilities of the SNPs in \code{case.genetic.data}. SNPs will be sampled in the
genetic algorithm proportional to the values specified. If not specified, by default, chi-square statistics of association will be computed for
each SNP, and sampling will be proportional to the root of these statistics. If user specified, the  vector values need not sum to 1, they just need to be positive
real numbers. See argument \code{prob} from function \code{sample} for more details.}

\item{categorical.exposures}{A matrix or data.frame of integers corresponding to categorical exposures for the cases. Defaults to NULL,
which will result in GADGETS looking for epistatic interactions, rather than SNP by exposure interactions. \code{categorical.exposures}
should not be missing any data, families with missing exposure data should be removed from the analysis prior to input.}

\item{continuous.exposures}{A matrix or data.frame of numeric values corresponding to continuous exposures for the cases. Defaults to NULL,
which will result in GADGETS looking for epistatic interactions, rather than SNP by exposure interactions. \code{continuous.exposures}
should not be missing any data, families with missing exposure data should be removed from the analysis prior to input.}

\item{categorical.exposures.sib}{If analyzing a study with siblings or a mix of trios and siblings, a matrix or data.frame of integers
corresponding to categorical exposures for the siblings and pseudo-siblings. Note that pseudo-siblings should be assigned the
same exposure as the case. Defaults to NULL. \code{categorical.exposures.sib}
should not be missing any data, families with missing exposure data should be removed from the analysis prior to input.}

\item{continuous.exposures.sib}{If analyzing a study with siblings or a mix of trios and siblings, a matrix or data.frame of numeric values
corresponding to continuous exposures for the siblings and pseudo-siblings. Note that pseudo-siblings should be assigned the same exposure as
the case. Defaults to NULL. code{continuous.exposures.sib} should not be missing any data, families with missing exposure data should be removed
from the analysis prior to input.}

\item{use.parents}{A integer indicating whether family level informativeness should be used alongside transmissions in computing GxE fitness scores. Defaults to 1,
indicating family level informativeness will be used. Specify 0 to only use transmission data.}

\item{mother.snps}{If searching for maternal-fetal interactions, the indices of the maternal SNPs in object 'case.genetic.data'. Otherwise does not need to be specified.}

\item{child.snps}{If searching for maternal-fetal interactions, the indices of the child SNPs in object 'case.genetic.data'. Otherwise does not need to be specified.}
}
\value{
A list containing the following:
\describe{
 \item{genetic.data.list}{A list of big.matrix.descriptor objects describing the locations of the input big.matrix objects
 containing the genetic data to be analyzed.}
 \item{chisq.stats}{A vector of chi-square statistics corresponding to marginal SNP-disease associations, if \code{snp.sampling.probs}
 is not specified, and \code{snp.sampling.probs} if specified.}
 \item{ld.block.vec}{A vector eaul to \code{cumsum(ld.block.vec)}.}
 \item{exposure}{A vector of categorical exposures, if specified, otherwise NULL.}
}
}
\description{
This function performs several pre-processing steps, intended for use before function run.gadgets.
}
\examples{

data(case)
data(dad)
data(mom)
case <- as.matrix(case)
dad <- as.matrix(dad)
mom <- as.matrix(mom)
res <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               ld.block.vec = c(10))

}
