% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.genetic.data.R
\name{preprocess.genetic.data}
\alias{preprocess.genetic.data}
\title{A function to pre-process case-parent triad of affected/unaffected sibling data.}
\usage{
preprocess.genetic.data(
  case.genetic.data,
  complement.genetic.data = NULL,
  father.genetic.data = NULL,
  mother.genetic.data = NULL,
  block.ld.mat,
  min.allele.freq = 0,
  bp.param = bpparam(),
  snp.sampling.probs = NULL,
  categorical.exposures = NULL
)
}
\arguments{
\item{case.genetic.data}{The genetic data of the disease affected children from case-parent trios or affected/unaffected sibling pairs. Columns are SNP allele counts, and rows are individuals.
The ordering of the columns must be consistent with the LD structure specified in \code{block.ld.mat}.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where
\code{complement.genetic.data} = mother SNP counts + father SNP counts - case SNP counts. If using affected/unaffected siblings
this should be the genotypes for the unaffected siblings.
Columns are SNP allele counts, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{father.genetic.data}{The genetic data for the fathers of the cases. Columns are SNP allele counts, rows are individuals.
Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{mother.genetic.data}{The genetic data for the mothers of the cases. Columns are SNP allele counts, rows are individuals.
Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{block.ld.mat}{A logical, block diagonal matrix indicating whether the SNPs in \code{case.genetic.data} should be considered
to be in linkage disequilibrium. Note that this means the ordering of the columns (SNPs) in \code{case.genetic.data} must be consistent
with the LD blocks specified in \code{ld.block.mat}. In the absence of outside information, a reasonable default is to consider SNPs
to be in LD if they are located on the same biological chromosome.}

\item{min.allele.freq}{The minimum minor allele frequency required for a SNP to be considered for inclusion in the genetic algorithm.
Any SNPs with MAF < \code{min.allele.freq} in the parents, or the combined group of affected and unaffected siblings, will be filtered out. Defaults to 0 (no filtering).}

\item{bp.param}{The BPPARAM argument to be passed to bplapply when estimating marginal disease associations for each SNP.
If using a cluster computer, this parameter needs to be set with care. See \code{BiocParallel::bplapply} for more details}

\item{snp.sampling.probs}{A vector indicating the sampling probabilities of the SNPs in \code{case.genetic.data}. SNPs will be sampled in the
genetic algorithm proportional to the values specified. If not specified, by default, chi-square statistics of association will be computed for
each SNP, and sampling will be proportional to the root of these statistics. If user specified, the  vector values need not sum to 1, they just need to be positive
real numbers. See argument \code{prob} from function \code{sample} for more details.}

\item{categorical.exposures}{A data.frame of categorical exposures for the cases. Rows correspond to individuals and columns correspond to categorical exposures.}
}
\value{
A list containing the following:
\describe{
 \item{case.genetic.data}{The pre-processed version of the case genetic data. Any missing genotypes for a given family will be coded as -9 for both case and complement,
 resulting in that family being uninformative for that SNP.}
 \item{complement.genetic.data}{Pre-processed complement or unaffected sibling genetic data. If mother and father data are input,
 the complement genetic data are first created and then pre-processed. Any missing genotypes for a given family will be coded as -9 for both case and complement,
 resulting in that family being uninformative for that SNP.}
 \item{chisq.stats}{A vector of chi-square statistics corresponding to marginal SNP-disease associations, if \code{snp.sampling.probs}
 is not specified, and \code{snp.sampling.probs} if specified.}
 \item{original.col.numbers}{A vector indicating the original column number of each non-filtered SNP remaining in the analysis data.}
 \item{block.ld.mat}{The pre-processed version of \code{block.ld.mat}.}
 \item{minor.allele.vec}{A vector indicating whether the alternate allele was the minor allele for each column in the input data.}
 \item{exposure}{A vector of categorical exposures, if specified, otherwise NULL.}
}
}
\description{
This function performs several pre-processing steps, intended for use before function run.gadgets.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
block.ld.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25),
                              matrix(rep(TRUE, 25^2), nrow = 25))))
res <- preprocess.genetic.data(case[, 1:10], father.genetic.data = dad[ , 1:10],
                               mother.genetic.data = mom[ , 1:10],
                               block.ld.mat = block.ld.mat[ , 1:10])

}
