% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ga.R
\name{run.ga}
\alias{run.ga}
\title{A function to run a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.}
\usage{
run.ga(
  case.genetic.data,
  complement.genetic.data = NULL,
  father.genetic.data = NULL,
  mother.genetic.data = NULL,
  n.chromosomes,
  chromosome.size,
  chrom.mat,
  seed.val,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  weight.function = identity,
  min.allele.freq = 0.025,
  generations = 2000,
  gen.same.fitness = 500,
  min.n.risk.set = 10,
  tol = 10^-6,
  n.top.chroms = 100,
  zscore.sd.threshold = 2.5,
  initial.sample.duplicates = F,
  snp.sampling.type = "zscore",
  warmup.gens = 0,
  crossover.prop = 0.5
)
}
\arguments{
\item{case.genetic.data}{A genetic dataset from cases (for a dichotomous trait). Columns are snps, and rows are individuals.}

\item{complement.genetic.data}{A genetic dataset from the complements of the cases, where \code{complement.genetic.data} = mother snp counts + father snp counts - case snp counts. Columns are snps, rows are families. If not specified, \code{father.genetic.data} and \code{mother.genetic.data} must be specified.}

\item{father.genetic.data}{The genetic data for the father of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{mother.genetic.data}{The genetic data for the mother of the case. Columns are snps, rows are individuals. Does not need to be specified if \code{complement.genetic.data} is specified.}

\item{n.chromosomes}{A scalar indicating the number of candidate collections of snps to use in the GA.}

\item{chromosome.size}{The number of snps within each candidate solution.}

\item{chrom.mat}{A logical matrix indicating whether the snps in the input genetic data belong to the same chromosome.}

\item{seed.val}{An integer indicating the seed to be used for the random samples.}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of different snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function}{A function that takes the weighted sum of the number of different snps and snps both equal to one as an argument, and returns a family weight. Defaults to the identity function.}

\item{min.allele.freq}{The minimum minor allele frequency in the parents required for a snp to be considered as a potential GA solution. Any snps with MAF < \code{min.allele.freq} in the parents will be omitted. Defaults to 0.01.}

\item{generations}{The maximum number of generations for which the GA will run. Defaults to 2000.}

\item{gen.same.fitness}{The number of consecutive generations with the same fitness score required for algorithm termination.}

\item{min.n.risk.set}{A scalar indicating the minimum number of individuals whose case - control difference vector must have sign consistent with the sign of the weighted sum of the differences vectors across families. Defaults to 10.}

\item{tol}{The maximum absolute pairwise difference among the top fitness scores from the previous 500 generations considered to be sufficient to stop producing new generations.}

\item{n.top.chroms}{The number of top scoring chromosomes, according to fitness score, to return.}

\item{zscore.sd.threshold}{A scalar indicating the maximum number of standard deviations from the mean a snp z-score can be. Any snps with marginal z-scores more extreme than this threshold will be set to the threshold value. This should be used to prevent very strong marginal associations from dominating the sampling.}

\item{initial.sample.duplicates}{A logical indicating whether the same snp can appear in more than one chromosome in the initial sample of chromosomes (the same snp may appear in more than one chromosome thereafter, regardless). Default to F.}

\item{snp.sampling.type}{A string indicating how snps are to be sampled for mutations. Options are "zscore" or "random". Defaults to "zscore".}

\item{warmup.gens}{A numeric indicating the number of warmup generations to be used. Defaults to 0.}

\item{crossover.prop}{A numeric between 0 and 1 indicating the proportion of chromosomes to be subjected to cross over. The remaining proportion will be mutated. Defaults to 0.5.}
}
\value{
A list, whose first element is a data.table of the top \code{n.top.chroms scoring chromosomes}, their fitness scores, and their difference vectors. The second element is a scalar indicating the number of generations required to identify a solution, and the third element is the number of snps filtered due to MAF < \code{min.allele.freq}.
}
\description{
This function runs a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500))))
ga.res <- run.ga(case, father.genetic.data = dad, mother.genetic.data = mom, n.chromosomes = 7,
                 chromosome.size = 3, chrom.mat = chrom.mat, seed.val = 10, generations = 1)

}
