% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ga.R
\name{run.ga}
\alias{run.ga}
\title{A function to run a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.}
\usage{
run.ga(
  data.list,
  n.chromosomes,
  chromosome.size,
  chrom.mat,
  seed.val,
  n.different.snps.weight = 2,
  n.both.one.weight = 1,
  weight.function = identity,
  generations = 2000,
  gen.same.fitness = 500,
  tol = 10^-6,
  n.top.chroms = 100,
  initial.sample.duplicates = F,
  snp.sampling.type = "chisq",
  crossover.prop = 0.5
)
}
\arguments{
\item{data.list}{The output list from \code{preprocess.genetic.data}.}

\item{n.chromosomes}{A scalar indicating the number of candidate collections of snps to use in the GA.}

\item{chromosome.size}{The number of snps within each candidate solution.}

\item{seed.val}{An integer indicating the seed to be used for the random samples.}

\item{n.different.snps.weight}{The number by which the number different snps between case and control is multiplied in computing the family weights. Defaults to 2.}

\item{n.both.one.weight}{The number by which the number of different snps equal to 1 in both case and control is multiplied in computing the family weights. Defaults to 1.}

\item{weight.function}{A function that takes the weighted sum of the number of different snps and snps both equal to one as an argument, and returns a family weight. Defaults to the identity function.}

\item{generations}{The maximum number of generations for which the GA will run. Defaults to 2000.}

\item{gen.same.fitness}{The number of consecutive generations with the same fitness score required for algorithm termination.}

\item{tol}{The maximum absolute pairwise difference among the top fitness scores from the previous 500 generations considered to be sufficient to stop producing new generations.}

\item{n.top.chroms}{The number of top scoring chromosomes, according to fitness score, to return.}

\item{initial.sample.duplicates}{A logical indicating whether the same snp can appear in more than one chromosome in the initial sample of chromosomes (the same snp may appear in more than one chromosome thereafter, regardless). Default to F.}

\item{snp.sampling.type}{A string indicating how snps are to be sampled for mutations. Options are "zscore" or "random". Defaults to "zscore".}

\item{crossover.prop}{A numeric between 0 and 1 indicating the proportion of chromosomes to be subjected to cross over. The remaining proportion will be mutated. Defaults to 0.5.}
}
\value{
A list, whose first element is a data.table of the top \code{n.top.chroms scoring chromosomes}, their fitness scores, and their difference vectors. The second element is a scalar indicating the number of generations required to identify a solution.
}
\description{
This function runs a genetic algorithm on a collection of candidate snp sets, to identify epistatic variants.
}
\examples{

data(case)
data(dad)
data(mom)
library(Matrix)
chrom.mat <- as.matrix(bdiag(list(matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500),
                              matrix(rep(TRUE, 2500^2), nrow = 2500))))
ga.res <- run.ga(case, father.genetic.data = dad, mother.genetic.data = mom, n.chromosomes = 7,
                 chromosome.size = 3, chrom.mat = chrom.mat, seed.val = 10, generations = 1)

}
